Support compiler comparison on Pacu.

+++ 01_Compiler/06_DoesItVectorize/test.cc
@@ -201,7 +201,7 @@ int main() {
   }
   printf("\n");
   printf("# -------");
-  for (int i=0; i<kernels.size(); ++i) {
+  for (size_t i=0; i<kernels.size(); ++i) {
     printf("  ---------------");
   }
   printf("\n");
